Compaund and a lot more stuff

+++ 015-the-wild-oasis/src/services/apiCabins.js
@@ -4,36 +4,47 @@ export async function getCabins() {
   const { data, error } = await supabase.from('cabins').select('*');
   if (error) {
     console.log(`!!! ${error} !!!`);
-    throw new Error(`📛 Cabins could not be loaded`);
+    throw new Error(`Cabins could not be loaded`);
   }
   return data;
 }
 
-// https://tqlpeeginxvocmwxjcur.supabase.co/storage/v1/object/public/cabin-images/cabin_001.jpg
-export async function createCabin(newCabin) {
-  const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll('/', '');
-  const imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+export async function createEditCabin(newCabin, id) {
+  const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
 
-  //--- 1. CREATE CABIN ---//
-  const { data, error } = await supabase
-    .from('cabins')
-    .insert([{ ...newCabin, image: imagePath }])
-    .select();
+  const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll('/', '');
+  const imagePath = hasImagePath
+    ? newCabin.image
+    : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+
+  // 1. Create/edit cabin
+  let query = supabase.from('cabins');
+
+  // A) CREATE
+  if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
+
+  // B) EDIT
+  if (id) query = query.update({ ...newCabin, image: imagePath }).eq('id', id);
+
+  const { data, error } = await query.select().single();
+
   if (error) {
-    console.log(`!!! ${error} !!!`);
-    throw new Error(`📛 Cabin could not be created`);
+    console.error(error);
+    throw new Error('Cabin could not be created');
   }
 
-  //--- 2. UPLOAD IMAGE ---//
+  // 2. Upload image
+  if (hasImagePath) return data;
+
   const { error: storageError } = await supabase.storage
     .from('cabin-images')
     .upload(imageName, newCabin.image);
 
-  //--- 3. DELETE THE CABIN IF THERE WAS AN UPLOADING ERROR ---//
+  // 3. Delete the cabin IF there was an error uplaoding image
   if (storageError) {
     await supabase.from('cabins').delete().eq('id', data.id);
-    console.log(`!!! ${storageError} !!!`);
-    throw new Error(`📛 Cabin image could not be uploaded and cabin could not be created`);
+    console.error(storageError);
+    throw new Error('Cabin image could not be uploaded and the cabin was not created');
   }
 
   return data;
@@ -43,7 +54,7 @@ export async function deleteCabin(id) {
   const { data, error } = await supabase.from('cabins').delete().eq('id', id);
   if (error) {
     console.log(`!!! ${error} !!!`);
-    throw new Error(`📛 Cabin could not be deleted`);
+    throw new Error(`Cabin could not be deleted`);
   }
   return data;
 }
